refactor: move fetch logic to helper.js and move global variables to config.js
