+++ src/js/model.js
@@ -1,4 +1,5 @@
-// import {async} from 'regenerator-runtime';
+import { API_URL } from './config.js';
+import { getJSON } from './helpers.js';
 
 export const state = {
   recipe: {},
@@ -6,12 +7,7 @@ export const state = {
 
 export const loadRecipe = async function (recipeId) {
   try {
-    const res = await fetch(
-      `https://forkify-api.herokuapp.com/api/v2/recipes/${recipeId}`
-    );
-    const data = await res.json();
-
-    if (!res.ok) throw new Error(`🚫 ${data.message} Error:${res.status} 🚫`);
+    const data = await getJSON(`${API_URL}/${recipeId}`);
 
     const {
       id,
@@ -35,6 +31,6 @@ export const loadRecipe = async function (recipeId) {
       ingredients,
     };
   } catch (err) {
-    console.error(err);
+    console.error(`${err} 🧑‍💻`);
   }
 };
