+++ src/js/views/recipeView.js
@@ -86,7 +86,11 @@ class RecipeView {
     <div class="recipe__ingredients">
       <h2 class="heading--2">Recipe ingredients</h2>
       <ul class="recipe__ingredient-list">
-      ${this.#data.ingredients.map(this.#generateMarkupIngredient).join(' ')}
+      ${
+        this.#data.ingredients
+          ? this.#data.ingredients.map(this.#generateMarkupIngredient).join(' ')
+          : ''
+      }
       </ul>
     </div>
